Push QLKS_27_5.bak Update Tinh tien theo ngay Update Tinh tien dich vu

+++ src/QLKS/DTO/ChiTietTienThuePhong.cs
@@ -16,7 +16,7 @@ namespace QLKS.DTO
 
         public double? Gia { get; set; }
 
-        public int SoLuong { get; set; }
+        public double SoLuong { get; set; }
 
         public double? ThanhTien { get; set; }
